Розділено контроллер AccountController на два різних для дотримання SRP
